edit WeatherMapApi
