+++ composer.json
@@ -11,7 +11,7 @@
         "php": "^7.2.5",
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^1.0",
-        "guzzlehttp/guzzle": "^6.3",
+        "guzzlehttp/guzzle": "7.0",
         "laravel/framework": "^7.0",
         "laravel/telescope": "^3.1",
         "laravel/tinker": "^2.0",
