+++ config/app.php
@@ -176,6 +176,7 @@ return [
         App\Providers\RouteServiceProvider::class,
         App\Providers\TelescopeServiceProvider::class,
         App\Providers\RolesServiceProvider::class,
+        App\Providers\WeatherMapApiServiceProvider::class,
 
     ],
